Separated the send email verification function for better re-use
